[Http] Moving Router classes to a new Routing\Dispatcher namespace for better organization.
