+++ src/Valkyrja/Http/Routing/Dispatcher/Contract/Router.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Http\Routing\Contract;
+namespace Valkyrja\Http\Routing\Dispatcher\Contract;
 
 use Valkyrja\Http\Message\Request\Contract\ServerRequest;
 use Valkyrja\Http\Message\Response\Contract\Response;
