+++ src/Valkyrja/Http/Routing/Provider/ServiceProvider.php
@@ -30,15 +30,15 @@ use Valkyrja\Http\Routing\Collection\Collection as HttpRoutingCollection;
 use Valkyrja\Http\Routing\Collection\Contract\Collection;
 use Valkyrja\Http\Routing\Collector\AttributeCollector;
 use Valkyrja\Http\Routing\Collector\Contract\Collector;
-use Valkyrja\Http\Routing\Contract\Router;
 use Valkyrja\Http\Routing\Data;
+use Valkyrja\Http\Routing\Dispatcher\Contract\Router;
+use Valkyrja\Http\Routing\Dispatcher\Router as HttpRouter;
 use Valkyrja\Http\Routing\Factory\Contract\ResponseFactory;
 use Valkyrja\Http\Routing\Matcher\Contract\Matcher;
 use Valkyrja\Http\Routing\Middleware\RequestStructMiddleware;
 use Valkyrja\Http\Routing\Middleware\ResponseStructMiddleware;
 use Valkyrja\Http\Routing\Middleware\ViewRouteNotMatchedMiddleware;
 use Valkyrja\Http\Routing\Processor\Contract\Processor;
-use Valkyrja\Http\Routing\Router as HttpRouter;
 use Valkyrja\Http\Routing\Url\Contract\Url;
 use Valkyrja\Reflection\Contract\Reflection;
 use Valkyrja\View\Renderer\Contract\Renderer;
