+++ tests/Unit/Http/Routing/Provider/ServiceProviderTest.php
@@ -36,8 +36,9 @@ use Valkyrja\Http\Routing\Collection\Collection;
 use Valkyrja\Http\Routing\Collection\Contract\Collection as CollectionContract;
 use Valkyrja\Http\Routing\Collector\AttributeCollector;
 use Valkyrja\Http\Routing\Collector\Contract\Collector as CollectorContract;
-use Valkyrja\Http\Routing\Contract\Router as RouterContract;
 use Valkyrja\Http\Routing\Data;
+use Valkyrja\Http\Routing\Dispatcher\Contract\Router as RouterContract;
+use Valkyrja\Http\Routing\Dispatcher\Router;
 use Valkyrja\Http\Routing\Factory\Contract\ResponseFactory as ResponseFactoryContract;
 use Valkyrja\Http\Routing\Factory\ResponseFactory;
 use Valkyrja\Http\Routing\Matcher\Contract\Matcher as MatcherContract;
@@ -48,7 +49,6 @@ use Valkyrja\Http\Routing\Middleware\ViewRouteNotMatchedMiddleware;
 use Valkyrja\Http\Routing\Processor\Contract\Processor as ProcessorContract;
 use Valkyrja\Http\Routing\Processor\Processor;
 use Valkyrja\Http\Routing\Provider\ServiceProvider;
-use Valkyrja\Http\Routing\Router;
 use Valkyrja\Http\Routing\Url\Contract\Url as UrlContract;
 use Valkyrja\Http\Routing\Url\Url;
 use Valkyrja\Reflection\Contract\Reflection;
