+++ tests/Unit/Http/Server/Psr/RequestHandlerTest.php
@@ -22,7 +22,7 @@ use Valkyrja\Http\Message\Response\Contract\Response as ResponseContract;
 use Valkyrja\Http\Message\Response\Response;
 use Valkyrja\Http\Message\Stream\Stream;
 use Valkyrja\Http\Middleware\Handler\RouteDispatchedHandler;
-use Valkyrja\Http\Routing\Router;
+use Valkyrja\Http\Routing\Dispatcher\Router;
 use Valkyrja\Http\Server\Psr\RequestHandler as PsrRequestHandler;
 use Valkyrja\Http\Server\RequestHandler;
 use Valkyrja\Tests\Unit\TestCase;
